+++ tests/IsAbsolutePathTest.php
@@ -13,8 +13,6 @@ class IsAbsolutePathTest extends TestCase
      */
     public function testIsAbsolutePath($path, $expected)
     {
-        require_once __DIR__.'/../src/command_functions.php';
-
         self::assertSame($expected, isAbsolutePath($path));
     }
